tests/wpt/tests/fedcm/fedcm-token-returned-with-http-error.https.html

<!DOCTYPE html>
<title>Federated Credential Management API token response tests</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {request_options_with_mediation_required,
        fedcm_error_dialog_click_button,
        fedcm_test,
        select_manifest,
        fedcm_get_and_select_first_account} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  const test_options =
  request_options_with_mediation_required("manifest_token_with_http_error.json");
  await select_manifest(t, test_options);

  const cred = fedcm_get_and_select_first_account(t, test_options);
  fedcm_error_dialog_click_button(t, "ErrorGotIt");
  try {
    await cred;
    assert_unreached("An IdentityCredentialError exception should be thrown.");
  } catch (e) {
    assert_true(e instanceof DOMException);
    assert_equals(e.name, "IdentityCredentialError");
    assert_equals(e.code, "");
    assert_equals(e.url, "");
  }
}, 'Test that the promise will be rejected if the response has http error');
</script>
